requirements.txt
annotated-types==0.7.0
anyio==4.6.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
fastapi==0.115.0
h11==0.14.0
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
numpy==2.1.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
python-multipart==0.0.12
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
sniffio==1.3.1
starlette==0.38.6
threadpoolctl==3.5.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.0